+++ CHANGELOG.md
@@ -11,5 +11,6 @@ All notable changes to this project are documented here. The format is based on
 -   Added RFC 2822 -> DateTime pipe.
 -   Added arbitrary format -> DateTime pipe.
 -   Added DateTime -> ISO 8601 pipes (full/date/time/week).
+-   Added DateTime -> artibrary format pipe.
 
 [keep-a-changelog]: http://keepachangelog.com/en/1.0.0/